Pretty much completed Block class I think

+++ mindblock/Colour.hpp
@@ -1,8 +1,10 @@
 #ifndef SAXBOPHONE_MINDBLOCK_COLOUR_HPP
 #define SAXBOPHONE_MINDBLOCK_COLOUR_HPP
 
+#include <cstdint>
+
 namespace mindblock {
-    enum struct Colour : uint {
+    enum struct Colour : uint_fast8_t {
         Grey, // default bland colour
         // the rainbow, but divided mathematically according to light primaries
         Red,
