+++ mindblock/Shape.hpp
@@ -1,8 +1,10 @@
 #ifndef SAXBOPHONE_MINDBLOCK_SHAPE_HPP
 #define SAXBOPHONE_MINDBLOCK_SHAPE_HPP
 
+#include <cstdint>
+
 namespace mindblock {
-    enum struct Shape : uint {
+    enum struct Shape : uint_fast8_t {
         Square,
     };
 }
